backend/requirements.txt
fastapi>=0.112.0
uvicorn[standard]>=0.30.0
python-dotenv>=1.0.1
psycopg[binary]
python-docx>=1.1.2
pydantic>=2.6.0
python-multipart>=0.0.9

# 管理系统依赖
bcrypt>=4.0.0
pyjwt>=2.8.0
email-validator>=2.0.0
